requirements.txt
mecab-python
